added galaxy overlap test, moved some functions around
I added a test that sees whether a galaxy is completely contained
within the virial radius of another. This is used in the halo finding
process to eliminate small subhalos.

I also moved some of the stuff from the KDE file (that weren’t in the
class) into the utils file where they really should have been in the
first place.