larger kernel size
I’m hoping this will fix some of the issues with centering. I’m
thinking they come from a kernel that is too small compared to the grid
size, so that if the real maximum locations aren’t right near a grid
point, they won’t be properly evaluated. This is basically a test to
see if that will change with a larger kernel.