proper kernel resizing
The kernel now is smaller for the particles in the center (no matter where we are sampling), rather then being smaller if the point of interest is in the center.

spherical KDE profile was removed in galaxy

radii were updated in the profile to be linear only.

KDE profiles can only be centered on zero (in any number of dimensions).

KDE locations use numpy arrays

1D gaussian uses same error checking

multiple sigmas can be used in a Gaussian kernel, which is needed for the different KDE kernels.