Added whole galaxy axis ratios
The radius within to calculate the disk plane is not passed into the initializing. It is also written to the output file for the full write.

I wrote a function to make the axis ratios object for the full galaxy.

Then the disk plane calculation is refactored to use this new function.